Moved access provider annotations to access.provider package
