+++ difuse-core/src/main/java/com/kaba4cow/difuse/core/annotation/access/provider/AllowPackages.java
@@ -1,4 +1,4 @@
-package com.kaba4cow.difuse.core.annotation.access;
+package com.kaba4cow.difuse.core.annotation.access.provider;
 
 import static java.lang.annotation.ElementType.METHOD;
 import static java.lang.annotation.ElementType.TYPE;
